+++ backend/src/Application/Modules/Accounts/Queries/GetOrganization.cs
@@ -4,6 +4,7 @@ using DarkDispatcher.Application.Modules.Accounts.Projections;
 using DarkDispatcher.Core.Persistence;
 using DarkDispatcher.Core.Queries;
 using DarkDispatcher.Domain.Accounts;
+using DarkDispatcher.Domain.Accounts.Ids;
 
 namespace DarkDispatcher.Application.Modules.Accounts.Queries;
