+++ backend/src/Application/Modules/Accounts/Queries/GetOrganizationsByUser.cs
@@ -5,6 +5,7 @@ using DarkDispatcher.Application.Modules.Accounts.Projections;
 using DarkDispatcher.Core.Persistence;
 using DarkDispatcher.Core.Queries;
 using DarkDispatcher.Domain.Accounts;
+using DarkDispatcher.Domain.Accounts.Ids;
 
 namespace DarkDispatcher.Application.Modules.Accounts.Queries;
