+++ backend/src/Domain/Accounts/Ids/OrganizationId.cs
@@ -1,0 +1,5 @@
+﻿using DarkDispatcher.Core.Aggregates;
+
+namespace DarkDispatcher.Domain.Accounts.Ids;
+
+public record OrganizationId(string Value) : AggregateId(Value);
