+++ backend/src/Domain/Features/Entities/Rule.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace DarkDispatcher.Domain.Features;
+namespace DarkDispatcher.Domain.Features.Entities;
 
 public record Rule(string Name, RuleServeType ServeType, string? Variation = null)
 {
