+++ backend/src/Domain/Features/Entities/RuleCondition.cs
@@ -1,3 +1,3 @@
-namespace DarkDispatcher.Domain.Features;
+namespace DarkDispatcher.Domain.Features.Entities;
 
 public record RuleCondition(RuleAttributeType Attribute, RuleComparer Comparer, string[] Values);
