+++ backend/src/Domain/Features/Entities/RuleVariationDefaults.cs
@@ -1,3 +1,3 @@
-﻿namespace DarkDispatcher.Domain.Features;
+﻿namespace DarkDispatcher.Domain.Features.Entities;
 
 public record RuleVariationDefaults(string On, string Off);
