+++ backend/src/Domain/Features/Entities/Variation.cs
@@ -1,3 +1,3 @@
-namespace DarkDispatcher.Domain.Features;
+namespace DarkDispatcher.Domain.Features.Entities;
 
 public record Variation(string Id, string Value, string? Name = null, string? Description = null);
