+++ backend/src/Domain/Features/Entities/VariationRollout.cs
@@ -1,3 +1,3 @@
-﻿namespace DarkDispatcher.Domain.Features;
+﻿namespace DarkDispatcher.Domain.Features.Entities;
 
 public record VariationRollout(string VariantId, int Weight);
