+++ backend/src/Domain/Features/Events/v1/FeatureUpdated.cs
@@ -1,0 +1,7 @@
+﻿using DarkDispatcher.Core.Events;
+using DarkDispatcher.Domain.Features.Ids;
+using DarkDispatcher.Domain.Projects.Entities;
+
+namespace DarkDispatcher.Domain.Features.Events.v1;
+
+public record FeatureUpdated(FeatureId FeatureId, string Name, Tag[]? Tags = null, string? Description = null) : DomainEvent;
