+++ backend/src/Domain/Features/Events/v1/VariationDeleted.cs
@@ -1,0 +1,6 @@
+﻿using DarkDispatcher.Core.Events;
+using DarkDispatcher.Domain.Features.Ids;
+
+namespace DarkDispatcher.Domain.Features.Events.v1;
+
+public record VariationDeleted(FeatureId FeatureId, string Value) : DomainEvent;
