+++ backend/src/Domain/Features/Events/v1/VariationUpdated.cs
@@ -1,0 +1,6 @@
+﻿using DarkDispatcher.Core.Events;
+using DarkDispatcher.Domain.Features.Ids;
+
+namespace DarkDispatcher.Domain.Features.Events.v1;
+
+public record VariationUpdated(FeatureId FeatureId, string Id, string Value, string? Name = null, string? Description = null) : DomainEvent;
