+++ backend/src/Domain/Features/Feature.cs
@@ -1,15 +1,15 @@
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using DarkDispatcher.Core.Aggregates;
 using DarkDispatcher.Core.Extensions;
+using DarkDispatcher.Domain.Features.Entities;
+using DarkDispatcher.Domain.Features.Events.v1;
+using DarkDispatcher.Domain.Features.Ids;
+using DarkDispatcher.Domain.Features.States;
 using DarkDispatcher.Domain.Projects;
-using DarkDispatcher.Domain.Projects.Ids;
-using DarkDispatcher.Domain.Projects.Models;
+using DarkDispatcher.Domain.Projects.Entities;
 
 namespace DarkDispatcher.Domain.Features;
-
-public record FeatureId(ConfigurationId ConfigurationId, string Value) : AggregateId(ConfigurationId.TenantId, Value);
 
 public sealed class Feature : Aggregate<FeatureState, FeatureId>
 {
@@ -26,13 +26,13 @@ public sealed class Feature : Aggregate<FeatureState, FeatureId>
     if (State.Type is VariationType.String or VariationType.Number && variations.HasDuplicates(x => x.Value)) 
       throw new InvalidOperationException("All variation values must be unique.");
 
-    var @event = new FeatureEvents.V1.FeatureCreated(id, key, name, type, variations, defaults, tags, description);
+    var @event = new FeatureCreated(id, key, name, type, variations, defaults, tags, description);
     Apply(@event);
   }
 
   public void Update(string name, Tag[]? tags = null, string? description = null)
   {
-    var @event = new FeatureEvents.V1.FeatureUpdated(GetAggregateId(), name, tags, description);
+    var @event = new FeatureUpdated(GetAggregateId(), name, tags, description);
     Apply(@event);
   }
 
@@ -44,7 +44,7 @@ public sealed class Feature : Aggregate<FeatureState, FeatureId>
     if (State.Variations.Any(x => x.Value.Equals(value, StringComparison.InvariantCultureIgnoreCase)))
       throw new InvalidOperationException("All variation values must be unique.");
 
-    var @event = new FeatureEvents.V1.VariationAdded(GetAggregateId(), id, value, name, description);
+    var @event = new VariationAdded(GetAggregateId(), id, value, name, description);
     Apply(@event);
   }
 
@@ -57,7 +57,7 @@ public sealed class Feature : Aggregate<FeatureState, FeatureId>
     if (State.Variations.Any(x => x.Id != id && x.Value.Equals(value, StringComparison.InvariantCultureIgnoreCase)))
       throw new InvalidOperationException("All variation values must be unique.");
 
-    var @event = new FeatureEvents.V1.VariationUpdated(GetAggregateId(), id, value, name, description);
+    var @event = new VariationUpdated(GetAggregateId(), id, value, name, description);
     Apply(@event);
   }
 
@@ -67,74 +67,7 @@ public sealed class Feature : Aggregate<FeatureState, FeatureId>
       throw new InvalidOperationException(
         $"Deleting variations for type '{nameof(VariationType.Boolean)}' is not allowed.");
 
-    var @event = new FeatureEvents.V1.VariationDeleted(GetAggregateId(), value);
+    var @event = new VariationDeleted(GetAggregateId(), value);
     Apply(@event);
   }
-}
-
-public record FeatureState : AggregateState<FeatureState, FeatureId>
-{
-  public FeatureState()
-  {
-    On<FeatureEvents.V1.FeatureCreated>((state, created) => state with
-    {
-      Id = created.FeatureId,
-      Key = created.Key,
-      Name = created.Name,
-      Description = created.Description,
-      Type = created.Type,
-      Variations = created.Variations.ToList(),
-      OnVariation = created.Defaults.On,
-      OffVariation = created.Defaults.Off
-    });
-
-    On<FeatureEvents.V1.FeatureUpdated>((state, updated) => state with
-    {
-      Name = updated.Name,
-      Description = updated.Description,
-      Tags = updated.Tags?.ToList() ?? new List<Tag>()
-    });
-
-    On<FeatureEvents.V1.VariationAdded>((state, added) =>
-    {
-      var variation = new Variation(added.Id, added.Value, added.Name, added.Description);
-      state.Variations.Add(variation);
-
-      return state;
-    });
-
-    On<FeatureEvents.V1.VariationUpdated>((state, updated) =>
-    {
-      var index = state.Variations.FindIndex(x => x.Id == updated.Id);
-      if (index >= 0)
-      {
-        state.Variations[index] = Variations[index] with
-        {
-          Value = updated.Value,
-          Name = updated.Name,
-          Description = updated.Description
-        };
-      }
-
-      return state;
-    });
-
-    On<FeatureEvents.V1.VariationDeleted>((state, deleted) =>
-    {
-      var variation = state.Variations.SingleOrDefault(x => x.Value == deleted.Value);
-      if (variation != null)
-        state.Variations.Remove(variation);
-
-      return state;
-    });
-  }
-
-  public string Key { get; init; } = null!;
-  public string Name { get; init; } = null!;
-  public string? Description { get; init; }
-  public VariationType Type { get; init; }
-  public string OnVariation { get; init; }
-  public string OffVariation { get; init; }
-  public List<Variation> Variations { get; init; } = new();
-  public List<Tag> Tags { get; init; } = new();
 }
