+++ backend/src/Domain/Projects/Entities/Tag.cs
@@ -1,3 +1,3 @@
-namespace DarkDispatcher.Domain.Projects.Models;
+namespace DarkDispatcher.Domain.Projects.Entities;
 
 public record Tag(string Id, string Name, TagColor Color);
