+++ backend/src/Domain/Projects/Ids/ProjectId.cs
@@ -1,5 +1,6 @@
 ﻿using DarkDispatcher.Core.Aggregates;
 using DarkDispatcher.Domain.Accounts;
+using DarkDispatcher.Domain.Accounts.Ids;
 
 namespace DarkDispatcher.Domain.Projects.Ids;
