+++ backend/src/Domain/Projects/Project.cs
@@ -1,7 +1,7 @@
 using DarkDispatcher.Core.Aggregates;
+using DarkDispatcher.Domain.Projects.Entities;
 using DarkDispatcher.Domain.Projects.Events.v1;
 using DarkDispatcher.Domain.Projects.Ids;
-using DarkDispatcher.Domain.Projects.Models;
 using DarkDispatcher.Domain.Projects.States;
 
 namespace DarkDispatcher.Domain.Projects;
